Switch Qwen3 model defaults to bf16 (fp16 causes overflow/NaN on Qwen)
Qwen3 models are trained natively in bf16. fp16 is known to cause
activation overflow in attention/MLP layers due to limited range
(max ~65504 vs bf16's ~3.4e38). bf16 also simplifies loading — model
loads directly in bf16, no GradScaler needed.

+++ train.py
@@ -131,7 +131,7 @@ MODEL_DEFAULTS = {
         "lr": 7e-5,
         "beta": 1e-3,
         "num_generations": 16,
-        "fp16": True,
+        "bf16": True,
         "gradient_checkpointing": True,
     },
     "Qwen3-4B": {
@@ -139,7 +139,7 @@ MODEL_DEFAULTS = {
         "lr": 7e-5,
         "beta": 1e-3,
         "num_generations": 16,
-        "fp16": True,
+        "bf16": True,
         "gradient_checkpointing": True,
     },
 }
